remove deprecated ToxDNS
Services & clients have moved to ToxMe.

Removal of ToxDNS from codebase is a clear sign that it is to not be
treated as standard and that no one should be using this.

No client and no service aside from those ran by µTox are using this.

This should help move µTox to the ToxMe.